add diff to CodeFeature based on intervals
